+++ Frends.Kafka.Produce/Frends.Kafka.Produce/Definitions/Enums.cs
@@ -114,3 +114,18 @@ public enum SslEndpointIdentificationAlgorithms
     Https
 #pragma warning restore CS1591 // self explanatory
 }
+
+/// <summary>
+/// Authentication credentials source.
+/// </summary>
+public enum AuthCredentialsSources
+{
+    /// <summary>
+    /// Credentials are specified via the `schema.registry.basic.auth.user.info` config property in the form username:password. If `schema.registry.basic.auth.user.info` is not set, authentication is disabled.
+    /// </summary>
+    UserInfo,
+    /// <summary>
+    /// Credentials are specified via the `sasl.username` and `sasl.password` configuration properties.
+    /// </summary>
+    SaslInherit
+}