+++ Frends.Kafka.Produce/Frends.Kafka.Produce/Definitions/Ssl.cs
@@ -37,45 +37,45 @@ public class Ssl
     /// <summary>
     /// Path to client's public key (PEM) used for authentication.
     /// </summary>
-    /// <example>c:\temp</example>
+    /// <example>/path/to/your/file</example>
     [UIHint(nameof(UseSsl), "", true)]
-    public string SslCertificateLocation { get; set; } = "";
+    public string SslCertificateLocation { get; set; }
 
     /// <summary>
     /// File or directory path to CA certificate(s) for verifying the broker's key.
     /// </summary>
     /// <example>Root</example>
     [UIHint(nameof(UseSsl), "", true)]
-    public string SslCaLocation { get; set; } = "";
+    public string SslCaLocation { get; set; }
 
     /// <summary>
     /// Path to client's private key (PEM) used for authentication.
     /// </summary>
-    /// <example>c:\temp</example>
+    /// <example>/path/to/your/file</example>
     [UIHint(nameof(UseSsl), "", true)]
-    public string SslKeyLocation { get; set; } = "";
+    public string SslKeyLocation { get; set; }
 
     /// <summary>
     /// Path to client's keystore (PKCS#12) used for authentication.
     /// </summary>
-    /// <example>c:\temp</example>
+    /// <example>/path/to/your/file</example>
     [UIHint(nameof(UseSsl), "", true)]
-    public string SslKeystoreLocation { get; set; } = "";
+    public string SslKeystoreLocation { get; set; }
 
     /// <summary>
     /// Path to OpenSSL engine library. 
     /// OpenSSL >= 1.1.0 required.
     /// </summary>
-    /// <example>c:\temp</example>
+    /// <example>/path/to/your/file</example>
     [UIHint(nameof(UseSsl), "", true)]
-    public string SslEngineLocation { get; set; } = "";
+    public string SslEngineLocation { get; set; }
 
     /// <summary>
     /// Path to CRL for verifying broker's certificate validity.
     /// </summary>
-    /// <example>c:\temp</example>
+    /// <example>/path/to/your/file</example>
     [UIHint(nameof(UseSsl), "", true)]
-    public string SslCrlLocation { get; set; } = "";
+    public string SslCrlLocation { get; set; }
 
     /// <summary>
     /// Client's public key string (PEM format) used for authentication.
@@ -83,7 +83,7 @@ public class Ssl
     /// <example>—–BEGIN PRIVATE KEY—–MIIES42Cg6zn—–END PRIVATE KEY—–</example>
     [UIHint(nameof(UseSsl), "", true)]
     [PasswordPropertyText]
-    public string SslCertificatePem { get; set; } = "";
+    public string SslCertificatePem { get; set; }
 
     /// <summary>
     /// A certificate string (PEM format) for verifying the broker's key.
@@ -91,7 +91,7 @@ public class Ssl
     /// <example>—–BEGIN PRIVATE KEY—–MIIES42Cg6zn—–END PRIVATE KEY—–</example>
     [UIHint(nameof(UseSsl), "", true)]
     [PasswordPropertyText]
-    public string SslCaPem { get; set; } = "";
+    public string SslCaPem { get; set; }
 
     /// <summary>
     /// Client's private key string (PEM format) used for authentication.
@@ -99,7 +99,7 @@ public class Ssl
     /// <example>—–BEGIN PRIVATE KEY—–MIIES42Cg6zn—–END PRIVATE KEY—–</example>
     [UIHint(nameof(UseSsl), "", true)]
     [PasswordPropertyText]
-    public string SslKeyPem { get; set; } = "";
+    public string SslKeyPem { get; set; }
 
     /// <summary>
     /// Comma-separated list of Windows Certificate stores to load CA certificates from. 
@@ -118,7 +118,7 @@ public class Ssl
     /// <example>ExamplePassword</example>
     [UIHint(nameof(UseSsl), "", true)]
     [PasswordPropertyText]
-    public string SslKeystorePassword { get; set; } = "";
+    public string SslKeystorePassword { get; set; }
 
     /// <summary>
     /// Private key passphrase for use with Ssl.SslKeyLocation)
@@ -126,26 +126,26 @@ public class Ssl
     /// <example>ExamplePassword</example>
     [UIHint(nameof(UseSsl), "", true)]
     [PasswordPropertyText]
-    public string SslKeyPassword { get; set; } = "";
+    public string SslKeyPassword { get; set; }
 
     /// <summary>
     /// A cipher suite is a named combination of authentication, encryption, MAC and key exchange algorithm used to negotiate the security settings for a network connection using TLS or SSL network protocol.
     /// </summary>
     /// <example>foo</example>
     [UIHint(nameof(UseSsl), "", true)]
-    public string SslCipherSuites { get; set; } = "";
+    public string SslCipherSuites { get; set; }
 
     /// <summary>
     /// The supported-curves extension in the TLS ClientHello message specifies the curves (standard/named, or 'explicit' GF(2^k) or GF(p)) the client is willing to have the server use.
     /// </summary>
     /// <example>1</example>
     [UIHint(nameof(UseSsl), "", true)]
-    public string SslCurvesList { get; set; } = "";
+    public string SslCurvesList { get; set; }
 
     /// <summary>
     /// The client uses the TLS ClientHello signature_algorithms extension to indicate to the server which signature/hash algorithm pairs may be used in digital signatures.
     /// </summary>
     /// <example>1</example>
     [UIHint(nameof(UseSsl), "", true)]
-    public string SslSigalgsList { get; set; } = "";
+    public string SslSigalgsList { get; set; }
 }